+++ communication/src/main/java/org/jboss/da/communication/cartographer/api/CartographerConnector.java
@@ -1,8 +1,8 @@
 package org.jboss.da.communication.cartographer.api;
 
 import org.jboss.da.common.CommunicationException;
-import org.jboss.da.communication.aprox.FindGAVDependencyException;
-import org.jboss.da.communication.aprox.model.GAVDependencyTree;
+import org.jboss.da.communication.indy.FindGAVDependencyException;
+import org.jboss.da.communication.indy.model.GAVDependencyTree;
 import org.jboss.da.model.rest.GAV;
 
 public interface CartographerConnector {
