+++ communication/src/main/java/org/jboss/da/communication/indy/FindGAVDependencyException.java
@@ -1,7 +1,7 @@
-package org.jboss.da.communication.aprox;
+package org.jboss.da.communication.indy;
 
 /**
- * This exception is thrown when the GAV we want to analyze for dependencies cannot be found on the Aprox server
+ * This exception is thrown when the GAV we want to analyze for dependencies cannot be found on the Indy server
  */
 public class FindGAVDependencyException extends Exception {
