+++ communication/src/main/java/org/jboss/da/communication/indy/api/IndyConnector.java
@@ -1,4 +1,4 @@
-package org.jboss.da.communication.aprox.api;
+package org.jboss.da.communication.indy.api;
 
 import java.io.InputStream;
 import java.util.List;
@@ -9,12 +9,12 @@ import org.jboss.da.communication.repository.api.RepositoryException;
 import org.jboss.da.model.rest.GA;
 import org.jboss.da.model.rest.GAV;
 
-public interface AproxConnector {
+public interface IndyConnector {
 
     /**
      * Finds available versions of specific groupId artifactId. If the provided groupId artifactId is not found in
      * repository, returns empty list.
-     * 
+     *
      * @param ga
      * @return list of versions for given groupId artifactId in repository.
      * @throws RepositoryException When there is problem with communication.
@@ -58,11 +58,11 @@ public interface AproxConnector {
     Optional<InputStream> getPomStream(GAV gav) throws RepositoryException;
 
     /**
-     * Finds out if a particular gav is present in the public repository listed by Aprox
-     * 
+     * Finds out if a particular gav is present in the public repository listed by Indy
+     *
      * @param gav
      * @return boolean
-     * @throws RepositoryException if we can't connect to the aprox server
+     * @throws RepositoryException if we can't connect to the Indy server
      */
     boolean doesGAVExistInPublicRepo(GAV gav) throws RepositoryException;
 }
