+++ communication/src/main/java/org/jboss/da/communication/indy/model/Versions.java
@@ -1,4 +1,4 @@
-package org.jboss.da.communication.aprox.model;
+package org.jboss.da.communication.indy.model;
 
 import java.util.List;
