+++ communication/src/main/java/org/jboss/da/communication/indy/model/npm/NpmMetadata.java
@@ -1,4 +1,4 @@
-package org.jboss.da.communication.aprox.model.npm;
+package org.jboss.da.communication.indy.model.npm;
 
 import java.util.Map;
