+++ communication/src/main/java/org/jboss/da/communication/pom/api/PomAnalyzer.java
@@ -9,7 +9,7 @@ import org.commonjava.maven.galley.maven.GalleyMavenException;
 import org.commonjava.maven.galley.maven.model.view.MavenPomView;
 import org.jboss.da.common.CommunicationException;
 import org.jboss.da.common.util.ConfigurationParseException;
-import org.jboss.da.communication.aprox.model.GAVDependencyTree;
+import org.jboss.da.communication.indy.model.GAVDependencyTree;
 import org.jboss.da.communication.pom.PomAnalysisException;
 import org.jboss.da.communication.pom.model.MavenProject;
 import org.jboss.da.model.rest.GA;
