+++ reports-backend/src/main/java/org/jboss/da/reports/backend/api/DependencyTreeGenerator.java
@@ -1,10 +1,9 @@
 package org.jboss.da.reports.backend.api;
 
-import org.jboss.da.communication.aprox.FindGAVDependencyException;
-
 import org.apache.maven.scm.ScmException;
 import org.jboss.da.common.CommunicationException;
-import org.jboss.da.communication.aprox.model.GAVDependencyTree;
+import org.jboss.da.communication.indy.FindGAVDependencyException;
+import org.jboss.da.communication.indy.model.GAVDependencyTree;
 import org.jboss.da.communication.pom.PomAnalysisException;
 import org.jboss.da.model.rest.GAV;
 import org.jboss.da.reports.model.api.SCMLocator;
