+++ reports-rest/src/main/java/org/jboss/da/rest/facade/ReportsFacade.java
@@ -2,7 +2,7 @@ package org.jboss.da.rest.facade;
 
 import org.apache.maven.scm.ScmException;
 import org.jboss.da.common.CommunicationException;
-import org.jboss.da.communication.aprox.FindGAVDependencyException;
+import org.jboss.da.communication.indy.FindGAVDependencyException;
 import org.jboss.da.communication.pom.PomAnalysisException;
 import org.jboss.da.reports.api.AlignmentReportModule;
 import org.jboss.da.reports.api.BuiltReportModule;
